Test Update
* change exam update not finished

* refactor upadate exam and fix some bugs

* fix all update methodes except dont use picture and contact

* fix 3 bug

* fix bug

* add log

* add log

* add log

+++ src/main/java/com/exam/examapp/service/impl/subject/SubjectServiceImpl.java
@@ -78,9 +78,11 @@ public class SubjectServiceImpl implements SubjectService {
 
         subject.setName(request.name());
         subject.setSupportMath(request.isSupportMath());
-        fileService.deleteFile(IMAGE_PATH, subject.getLogoUrl());
-        String logoUrl = fileService.uploadFile(IMAGE_PATH, logo);
-        subject.setLogoUrl(logoUrl);
+        if (logo != null) {
+            fileService.deleteFile(IMAGE_PATH, subject.getLogoUrl());
+            String logoUrl = fileService.uploadFile(IMAGE_PATH, logo);
+            subject.setLogoUrl(logoUrl);
+        }
         subjectRepository.save(subject);
         log.info("Mövzu yeniləndi");
         logService.save("Mövzu yeniləndi", userService.getCurrentUserOrNull());
